Code re-factoring
