+++ src/defaults/DefaultLinkWidget.tsx
@@ -2,10 +2,11 @@ import * as React from "react";
 import * as _ from "lodash";
 import * as PF from "pathfinding";
 import * as Path from "paths-js/path";
-import { DiagramEngine, ROUTING_SCALING_FACTOR } from "../DiagramEngine";
+import { DiagramEngine } from "../DiagramEngine";
 import { LinkModel } from "../models/LinkModel";
 import { PointModel } from "../models/PointModel";
 import { NodeModel } from "../models/NodeModel";
+import PathFinding, { ROUTING_SCALING_FACTOR } from "../routing/PathFinding";
 
 export interface DefaultLinkProps {
 	color?: string;
@@ -19,11 +20,6 @@ export interface DefaultLinkProps {
 export interface DefaultLinkState {
 	selected: boolean;
 }
-
-const pathFinderInstance = new PF.JumpPointFinder({
-	heuristic: PF.Heuristic.manhattan,
-	diagonalMovement: PF.DiagonalMovement.Never
-});
 
 /**
  * @author Dylan Vorster
@@ -42,11 +38,17 @@ export class DefaultLinkWidget extends React.Component<DefaultLinkProps, Default
 	refLabel: HTMLElement;
 	refPaths: SVGPathElement[];
 
+	pathFinding: PathFinding; // only set when smart routing is active
+
 	constructor(props: DefaultLinkProps) {
 		super(props);
 		this.state = {
 			selected: false
 		};
+
+		if (props.diagramEngine.isSmartRoutingEnabled()) {
+			this.pathFinding = new PathFinding(this.props.diagramEngine);
+		}
 	}
 
 	addPointToLink = (event: MouseEvent, index: number): void => {
@@ -227,76 +229,7 @@ export class DefaultLinkWidget extends React.Component<DefaultLinkProps, Default
 		return path.print();
 	}
 
-	calculateDirectPath(
-		from: {
-			x: number;
-			y: number;
-		},
-		to: {
-			x: number;
-			y: number;
-		}
-	): number[][] {
-		const { diagramEngine } = this.props;
-		const matrix = diagramEngine.getCanvasMatrix();
-
-		const grid = new PF.Grid(matrix);
-
-		return pathFinderInstance.findPath(
-			diagramEngine.translateRoutingX(Math.floor(from.x / ROUTING_SCALING_FACTOR)),
-			diagramEngine.translateRoutingY(Math.floor(from.y / ROUTING_SCALING_FACTOR)),
-			diagramEngine.translateRoutingX(Math.floor(to.x / ROUTING_SCALING_FACTOR)),
-			diagramEngine.translateRoutingY(Math.floor(to.y / ROUTING_SCALING_FACTOR)),
-			grid
-		);
-	}
-
-	calculateLinkStartEndCoords(
-		matrix: number[][],
-		path: number[][]
-	): {
-		start: {
-			x: number;
-			y: number;
-		};
-		end: {
-			x: number;
-			y: number;
-		};
-		pathToStart: number[][];
-		pathToEnd: number[][];
-	} {
-		const startIndex = path.findIndex(point => matrix[point[1]][point[0]] === 0);
-		const endIndex =
-			path.length -
-			1 -
-			path
-				.slice()
-				.reverse()
-				.findIndex(point => matrix[point[1]][point[0]] === 0);
-
-		// are we trying to create a path exclusively through blocked areas?
-		// if so, let's fallback to the linear routing
-		if (startIndex === -1 || endIndex === -1) {
-			return undefined;
-		}
-
-		const pathToStart = path.slice(0, startIndex);
-		const pathToEnd = path.slice(endIndex);
-
-		return {
-			start: {
-				x: path[startIndex][0],
-				y: path[startIndex][1]
-			},
-			end: {
-				x: path[endIndex][0],
-				y: path[endIndex][1]
-			},
-			pathToStart,
-			pathToEnd
-		};
-	}
+	//HERE
 
 	/**
 	 * Smart routing is only applicable when all conditions below are true:
@@ -336,28 +269,19 @@ export class DefaultLinkWidget extends React.Component<DefaultLinkProps, Default
 
 		if (this.isSmartRoutingApplicable()) {
 			// first step: calculate a direct path between the points being linked
-			const directPathCoords = this.calculateDirectPath(_.first(points), _.last(points));
+			const directPathCoords = this.pathFinding.calculateDirectPath(_.first(points), _.last(points));
 
 			const routingMatrix = diagramEngine.getRoutingMatrix();
 			// now we need to extract, from the routing matrix, the very first walkable points
 			// so they can be used as origin and destination of the link to be created
-			const smartLink = this.calculateLinkStartEndCoords(routingMatrix, directPathCoords);
+			const smartLink = this.pathFinding.calculateLinkStartEndCoords(routingMatrix, directPathCoords);
 
 			if (smartLink) {
 				const { start, end, pathToStart, pathToEnd } = smartLink;
-				// second step: calculate a path avoiding hitting other elements
-				const grid = new PF.Grid(routingMatrix);
-				const dynamicPath = pathFinderInstance.findPath(start.x, start.y, end.x, end.y, grid);
 
-				// third step: aggregate everything to have the calculated path ready for render
-				const pathCoords = pathToStart
-					.concat(dynamicPath, pathToEnd)
-					.map(coords => [
-						diagramEngine.translateRoutingX(coords[0], true),
-						diagramEngine.translateRoutingY(coords[1], true)
-					]);
-				// const svgPath = this.generateDynamicPath(PF.Util.smoothenPath(grid, pathCoords));
-				const svgPath = this.generateDynamicPath(PF.Util.compressPath(pathCoords));
+				// second step: calculate a path avoiding hitting other elements
+				const simplifiedPath = this.pathFinding.calculateDynamicPath(routingMatrix, start, end, pathToStart, pathToEnd);
+				const svgPath = this.generateDynamicPath(simplifiedPath);
 
 				paths.push(
 					this.generateLink(
